+++ src/BloomExe/Edit/ConfirmRemovePageDialog.designer.cs
@@ -34,7 +34,7 @@
 			this.pictureRecycleBin = new System.Windows.Forms.PictureBox();
 			this.cancelBtn = new System.Windows.Forms.Button();
 			this.deleteBtn = new System.Windows.Forms.Button();
-			this.localizationExtender1 = new Localization.UI.LocalizationExtender(this.components);
+			this.localizationExtender1 = new L10NSharp.UI.LocalizationExtender(this.components);
 			this.tableLayout.SuspendLayout();
 			((System.ComponentModel.ISupportInitialize)(this.pictureRecycleBin)).BeginInit();
 			((System.ComponentModel.ISupportInitialize)(this.localizationExtender1)).BeginInit();
@@ -169,6 +169,6 @@
 		private System.Windows.Forms.PictureBox pictureRecycleBin;
 		private System.Windows.Forms.Button cancelBtn;
 		private System.Windows.Forms.Button deleteBtn;
-		private Localization.UI.LocalizationExtender localizationExtender1;
+		private L10NSharp.UI.LocalizationExtender localizationExtender1;
 	}
 }